Added LeafNode class, from which Heading and Paragraph extends. Deduplicates code, makes testing for Header or Paragraph (e.g. leaf node) simpler.
