+++ docs/jsdoc-conf.json
@@ -1,0 +1,3 @@
+{
+  "plugins": [ "plugins/markdown" ]
+}
