+++ src/tree/structure/nodes/index.js
@@ -1,4 +1,5 @@
 import Node from "./Node";
+import LeafNode from "./LeafNode";
 import List from "./List";
 import ListItem from "./ListItem";
 import Paragraph from "./Paragraph";
@@ -9,6 +10,7 @@ import Whitespace from "./Whitespace";
 
 export {
 	Node,
+	LeafNode,
 	StructuredNode,
 	Paragraph,
 	Heading,
